refactor: register logic to useRegisterUser hook
